minor fixes for message deletion

+++ src/data/tasks/remove_old_message_4.py
@@ -11,10 +11,12 @@ class Task_RemoveOldMessage(TaskBase):
     async def execute(cl, obj: object) -> None:
         if obj and obj["guild"] and obj["message_id"]:
             message_data = bot.instance.data.guilds.get(obj["guild"]).messages.get_by_message_id(obj["message_id"])
-            channel = bot.instance.get_channel(message_data.channel_id)
-            if channel:
-                message = await cache.messages.get(message_data.message_id, channel)
-                if message:
-                    await message.delete()
+            if message_data:
+                channel = bot.instance.get_channel(message_data.channel_id)
+                if channel:
+                    message = await cache.messages.get(message_data.message_id, channel)
+                    if message:
+                        await message.delete()
+                bot.instance.data.guilds.get(obj["guild"]).messages.remove(message_data.message_id)
